# examples/petstore3-server-go-echo/server/openapi_generated.yml
openapi: 3.0.2
info:
    contact:
        email: apiteam@swagger.io
    description: |-
        This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
        Swagger at [http://swagger.io](http://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
        You can now help us improve the API whether it's by making changes to the definition itself or to the code.
        That way, with time, we can improve the API in general, and expose some of the new features in OAS3.
    
        Some useful links:
        - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
        - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    termsOfService: http://swagger.io/terms/
    title: Swagger Petstore - OpenAPI 3.0
    version: 1.0.19
...
servers:
    - url: /api/v3
...
paths:
...
    /store/inventory:
        get:
            description: Returns a map of status codes to quantities
            operationId: getInventory
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                additionalProperties:
                                    format: int32
                                    type: integer
                                type: object
                    description: successful operation
            security:
                - api_key: []
            summary: Returns pet inventories by status
            tags:
                - store
    /store/order:
        post:
            description: Place a new order in the store
            operationId: placeOrder
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Order'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/Order'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/Order'
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Order'
                    description: successful operation
                "405":
                    description: Invalid input
            summary: Place an order for a pet
            tags:
                - store
    /store/order/{orderId}:
        get:
            description: For valid response try integer IDs with value <= 5 or > 10. Other values will generate exceptions.
            operationId: getOrderById
            parameters:
                - description: ID of order that needs to be fetched
                  in: path
                  name: orderId
                  required: true
                  schema:
                    format: int64
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Order'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Order'
                    description: successful operation
                "400":
                    description: Invalid ID supplied
                "404":
                    description: Order not found
            summary: Find purchase order by ID
            tags:
                - store
        delete:
            description: For valid response try integer IDs with value < 1000. Anything above 1000 or nonintegers will generate API errors
            operationId: deleteOrder
            parameters:
                - description: ID of the order that needs to be deleted
                  in: path
                  name: orderId
                  required: true
                  schema:
                    format: int64
                    type: integer
            responses:
                "400":
                    description: Invalid ID supplied
                "404":
                    description: Order not found
            summary: Delete purchase order by ID
            tags:
                - store
components:
...
    schemas:
...
        Order:
            properties:
                complete:
                    type: boolean
                id:
                    example: 10
                    format: int64
                    type: integer
                petId:
                    example: 198772
                    format: int64
                    type: integer
                quantity:
                    example: 7
                    format: int32
                    type: integer
                shipDate:
                    format: date-time
                    type: string
                status:
                    description: Order Status
                    enum:
                        - placed
                        - approved
                        - delivered
                    example: approved
                    type: string
            type: object
            xml:
                name: order
...
    securitySchemes:
        api_key:
            in: header
            name: api_key
            type: apiKey
        petstore_auth:
            flows:
                implicit:
                    authorizationUrl: https://petstore3.swagger.io/oauth/authorize
                    scopes:
                        read:pets: read your pets
                        write:pets: modify pets in your account
            type: oauth2